[Study programme] Limit year selector visibility
